Implement SPI traits for u16, u32 SPI peripherals
Add embedded_hal SPI trait implementations for u16 and u32 word
sizes. This lets us mock SPI peripherals for devices that expect
u16 or u32 word sizes.

We use a macro in the implementation. It seems that we cannot be
generic over the word size, since there's a default trait in the
embedded_hal crate that we might conflict. When trying to remain
generic over the word size, we find

    error[E0119]: conflicting implementations of trait `embedded_hal::blocking::spi::Transfer<_>` for type `common::Generic<spi::Transaction<_>>`:
    --> src/spi.rs:170:1
        |
    170 | impl<W> spi::Transfer<W> for Mock<W> {
        | ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
        |
        = note: conflicting implementation in crate `embedded_hal`:
                - impl<W, S> embedded_hal::blocking::spi::Transfer<W> for S
                where S: embedded_hal::blocking::spi::transfer::Default<W>, W: std::clone::Clone;
        = note: downstream crates may implement trait `embedded_hal::blocking::spi::transfer::Default<_>` for type `common::Generic<spi::Transaction<_>>`

If anyone knows a way to work around this, let me know!

This might break embedded_hal_mock users. We cannot rely on hard-coded
integers coercing to a u8 anymore. Users might be required to supply
the word type in their SPI mock. See the updated SPI documentation,
and the tests, for what was required to make the tests build after
this change.